Refactored User Identity lookups into single calss and cleaned up imports
