+++ src/main/java/io/boomerang/controller/TeamController.java
@@ -28,8 +28,8 @@ import io.boomerang.model.profile.SortSummary;
 import io.boomerang.model.teams.ApproverGroupResponse;
 import io.boomerang.model.teams.CreateApproverGroupRequest;
 import io.boomerang.mongo.entity.FlowTeamConfiguration;
-import io.boomerang.mongo.entity.TeamEntity;
 import io.boomerang.mongo.entity.FlowUserEntity;
+import io.boomerang.mongo.entity.TeamEntity;
 import io.boomerang.mongo.model.Quotas;
 import io.boomerang.mongo.model.UserType;
 import io.boomerang.service.UserIdentityService;
