+++ src/main/java/io/boomerang/controller/api/WorkflowsV1Controller.java
@@ -12,8 +12,8 @@ import io.boomerang.security.interceptors.AuthenticationScope;
 import io.boomerang.service.crud.WorkflowService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
 @RestController
