+++ src/main/java/io/boomerang/model/WorkflowScheduleCalendar.java
@@ -1,7 +1,6 @@
 package io.boomerang.model;
 
 import java.util.Date;
-
 /*
  * Maps a list of Dates returned by Quartz utils to a schedule for displaying on the calendar
  */
