+++ src/main/java/io/boomerang/model/controller/Workflow.java
@@ -2,7 +2,6 @@ package io.boomerang.model.controller;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
