+++ src/main/java/io/boomerang/mongo/entity/ActivityEntity.java
@@ -6,11 +6,11 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import io.boomerang.model.controller.TaskWorkspace;
-import io.boomerang.mongo.model.KeyValuePair;
 import io.boomerang.mongo.model.ErrorResponse;
+import io.boomerang.mongo.model.KeyValuePair;
 import io.boomerang.mongo.model.TaskStatus;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
