+++ src/main/java/io/boomerang/security/interceptors/AuthenticationScope.java
@@ -1,8 +1,9 @@
 package io.boomerang.security.interceptors;
 
+import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import java.lang.annotation.*;
 import io.boomerang.mongo.model.TokenScope;
 
 @Target({ElementType.METHOD, ElementType.TYPE})
