+++ src/main/java/io/boomerang/security/interceptors/SecurityConfig.java
@@ -4,13 +4,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import io.boomerang.security.service.UserDetailsService;
+import io.boomerang.service.UserIdentityService;
 
 @Configuration
 public class SecurityConfig implements WebMvcConfigurer {
   
   @Autowired
-  private UserDetailsService userDetailsService;
+  private UserIdentityService userDetailsService;
   
   @Override
   public void addInterceptors(InterceptorRegistry registry) {
