--- src/main/java/io/boomerang/security/service/UserDetailsService.java
@@ -1,9 +1,0 @@
-package io.boomerang.security.service;
-
-import io.boomerang.mongo.model.TokenScope;
-import io.boomerang.security.model.UserToken;
-
-public interface UserDetailsService {
-  public UserToken getUserDetails();
-  public TokenScope getCurrentScope();
-}
