+++ src/main/java/io/boomerang/service/FlowExecutionServiceImpl.java
@@ -28,13 +28,13 @@ import io.boomerang.mongo.entity.FlowTaskTemplateEntity;
 import io.boomerang.mongo.entity.RevisionEntity;
 import io.boomerang.mongo.entity.TaskExecutionEntity;
 import io.boomerang.mongo.entity.WorkflowEntity;
-import io.boomerang.mongo.model.KeyValuePair;
+import io.boomerang.mongo.model.ActivityStorage;
 import io.boomerang.mongo.model.Dag;
+import io.boomerang.mongo.model.KeyValuePair;
 import io.boomerang.mongo.model.Revision;
 import io.boomerang.mongo.model.Storage;
 import io.boomerang.mongo.model.TaskStatus;
 import io.boomerang.mongo.model.TaskType;
-import io.boomerang.mongo.model.ActivityStorage;
 import io.boomerang.mongo.model.internal.InternalTaskRequest;
 import io.boomerang.mongo.model.next.DAGTask;
 import io.boomerang.mongo.model.next.Dependency;
