+++ src/main/java/io/boomerang/service/crud/TeamServiceImpl.java
@@ -18,8 +18,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.HttpClientErrorException;
-import io.boomerang.client.ExternalUserService;
 import io.boomerang.client.ExternalTeamService;
+import io.boomerang.client.ExternalUserService;
 import io.boomerang.client.model.Team;
 import io.boomerang.client.model.UserProfile;
 import io.boomerang.model.FlowTeam;
@@ -35,16 +35,16 @@ import io.boomerang.model.teams.ApproverUser;
 import io.boomerang.model.teams.CreateApproverGroupRequest;
 import io.boomerang.mongo.entity.ActivityEntity;
 import io.boomerang.mongo.entity.FlowTeamConfiguration;
-import io.boomerang.mongo.entity.TeamEntity;
 import io.boomerang.mongo.entity.FlowUserEntity;
+import io.boomerang.mongo.entity.TeamEntity;
 import io.boomerang.mongo.entity.WorkflowEntity;
+import io.boomerang.mongo.model.ActivityStorage;
 import io.boomerang.mongo.model.ApproverGroup;
 import io.boomerang.mongo.model.Quotas;
 import io.boomerang.mongo.model.Settings;
 import io.boomerang.mongo.model.Storage;
 import io.boomerang.mongo.model.TaskStatus;
 import io.boomerang.mongo.model.UserType;
-import io.boomerang.mongo.model.ActivityStorage;
 import io.boomerang.mongo.service.FlowSettingsService;
 import io.boomerang.mongo.service.FlowTeamService;
 import io.boomerang.mongo.service.FlowUserService;
