+++ src/main/java/io/boomerang/service/crud/WorkflowScheduleServiceImpl.java
@@ -25,8 +25,8 @@ import io.boomerang.mongo.entity.WorkflowScheduleEntity;
 import io.boomerang.mongo.model.KeyValuePair;
 import io.boomerang.mongo.model.WorkflowScheduleStatus;
 import io.boomerang.mongo.model.WorkflowScheduleType;
-import io.boomerang.mongo.service.ScheduleService;
 import io.boomerang.mongo.service.FlowWorkflowService;
+import io.boomerang.mongo.service.ScheduleService;
 import io.boomerang.quartz.QuartzSchedulerService;
 import io.boomerang.service.FilterService;
 import io.boomerang.util.ParameterMapper;
